+++ src/DotNext.IO/IO/StreamExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System.Runtime.CompilerServices;
+using Microsoft.Win32.SafeHandles;
 
 namespace DotNext.IO;
 
@@ -25,4 +26,24 @@ public static partial class StreamExtensions
     /// <returns>An object that represents multiple streams as one logical stream.</returns>
     public static Stream Combine(this Stream stream, ReadOnlySpan<Stream> others)
         => others is { Length: > 0 } ? new SparseStream([stream, .. others]) : stream;
+
+    /// <summary>
+    /// Creates a stream for the specified file handle.
+    /// </summary>
+    /// <remarks>
+    /// The returned stream doesn't own the handle.
+    /// </remarks>
+    /// <param name="handle">The file handle.</param>
+    /// <param name="access">Desired access to the file via stream.</param>
+    /// <returns>The unbuffered file stream.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="handle"/> is <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentException"><paramref name="handle"/> is closed or invalid.</exception>
+    public static Stream AsUnbufferedStream(this SafeFileHandle handle, FileAccess access)
+    {
+        ArgumentNullException.ThrowIfNull(handle);
+
+        return handle is { IsInvalid: false, IsClosed: false }
+            ? new UnbufferedFileStream(handle, access)
+            : throw new ArgumentException(ExceptionMessages.FileHandleClosed, nameof(handle));
+    }
 }