+++ src/cluster/DotNext.Net.Cluster/Net/Cluster/Consensus/Raft/LeaderState.cs
@@ -11,14 +11,14 @@ using IO.Log;
 using Membership;
 using Runtime.CompilerServices;
 using Threading.Tasks;
-using static Threading.LinkedTokenSourceFactory;
+using GCLatencyModeScope = Runtime.GCLatencyModeScope;
 
 internal sealed partial class LeaderState<TMember> : RaftState<TMember>
     where TMember : class, IRaftClusterMember
 {
     private readonly long currentTerm;
     private readonly CancellationTokenSource timerCancellation;
-    internal readonly CancellationToken LeadershipToken; // cached to avoid ObjectDisposedException
+    internal readonly CancellationToken LeadershipToken; // cached to prevent ObjectDisposedException
     private readonly Task<Result<bool>> localMemberResponse;
 
     private Task? heartbeatTask;
@@ -38,7 +38,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
         replicatorFactory = localReplicatorFactory = CreateDefaultReplicator;
     }
 
-    private (long, long, int) ForkHeartbeats(TaskCompletionPipe<Task<Result<bool>>> responsePipe, IAuditTrail<IRaftLogEntry> auditTrail, IClusterConfigurationStorage configurationStorage, IEnumerator<TMember> members, CancellationToken token)
+    private (long, long, int) ForkHeartbeats(TaskCompletionPipe<Task<Result<bool>>> responsePipe, IAuditTrail<IRaftLogEntry> auditTrail, IClusterConfigurationStorage configurationStorage, IEnumerator<TMember> members)
     {
         Replicator replicator;
         Task<Result<bool>> response;
@@ -48,7 +48,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
         var majority = 0;
 
         // send heartbeat in parallel
-        for (IClusterConfiguration? activeConfig = configurationStorage.ActiveConfiguration, proposedConfig = configurationStorage.ProposedConfiguration; members.MoveNext(); responsePipe.Add(response, replicator), majority++)
+        for (IClusterConfiguration? activeConfig = configurationStorage.ActiveConfiguration, proposedConfig = configurationStorage.ProposedConfiguration; members.MoveNext() && !LeadershipToken.IsCancellationRequested; responsePipe.Add(response, replicator), majority++)
         {
             var member = members.Current;
             if (member.IsRemote)
@@ -58,7 +58,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
                 // fork replication procedure
                 replicator = context.GetOrCreate(member, replicatorFactory);
                 replicator.Initialize(activeConfig, proposedConfig, commitIndex, currentTerm, precedingIndex);
-                response = SpawnReplicationAsync(replicator, auditTrail, currentIndex, token);
+                response = SpawnReplicationAsync(replicator, auditTrail, currentIndex, LeadershipToken);
             }
             else
             {
@@ -80,7 +80,9 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
         {
             result = response.GetAwaiter().GetResult();
             detector?.ReportHeartbeat();
-            return currentTerm >= result.Term ? MemberResponse.Successful : MemberResponse.HigherTermDetected;
+            return currentTerm >= result.Term
+                ? MemberResponse.Successful
+                : MemberResponse.HigherTermDetected;
         }
         catch (MemberUnavailableException)
         {
@@ -120,45 +122,33 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
     }
 
     [AsyncMethodBuilder(typeof(SpawningAsyncTaskMethodBuilder))]
-    private async Task DoHeartbeats(TimeSpan period, IAuditTrail<IRaftLogEntry> auditTrail, IClusterConfigurationStorage configurationStorage, IReadOnlyCollection<TMember> members, CancellationToken token)
+    private async Task DoHeartbeats(TimeSpan period, IAuditTrail<IRaftLogEntry> auditTrail, IClusterConfigurationStorage configurationStorage, IReadOnlyCollection<TMember> members)
     {
-        var cancellationSource = token.LinkTo(LeadershipToken);
-
         // cached enumerator allows to avoid memory allocation on every GetEnumerator call inside of the loop
         var enumerator = members.GetEnumerator();
         try
         {
             var forced = false;
-            for (var responsePipe = new TaskCompletionPipe<Task<Result<bool>>>(); !token.IsCancellationRequested; responsePipe.Reset(), ReuseEnumerator(ref members, ref enumerator))
+            for (var responsePipe = new TaskCompletionPipe<Task<Result<bool>>>(); !LeadershipToken.IsCancellationRequested; responsePipe.Reset(), ReuseEnumerator(ref members, ref enumerator))
             {
                 var startTime = new Timestamp();
 
                 // do not resume suspended callers that came after the barrier, resume them in the next iteration
                 replicationQueue.SwitchValve();
 
-                GCLatencyMode latencyMode;
-                if (forced)
-                {
-                    // in case of forced (initiated programmatically, not by timeout) replication
-                    // do not change GC latency. Otherwise, in case of high load GC is not able to collect garbage
-                    Unsafe.SkipInit(out latencyMode);
-                }
-                else
-                {
-                    // we want to minimize GC intrusion during replication process
-                    // (however, it is still allowed in case of system-wide memory pressure, e.g. due to container limits)
-                    latencyMode = GCSettings.LatencyMode;
-                    GCSettings.LatencyMode = GCLatencyMode.SustainedLowLatency;
-                }
-
+                // in case of forced (initiated programmatically, not by timeout) replication
+                // do not change GC latency. Otherwise, in case of high load GC is not able to collect garbage
+                var latencyScope = forced
+                    ? default
+                    : GCLatencyModeScope.SustainedLowLatency;
                 try
                 {
                     // Perf: the code in this block is inlined instead of moved to separated method because
                     // we want to prevent allocation of state machine on every call
                     int quorum = 0, commitQuorum = 0, majority;
-                    (long currentIndex, long commitIndex, majority) = ForkHeartbeats(responsePipe, auditTrail, configurationStorage, enumerator, token);
+                    (long currentIndex, long commitIndex, majority) = ForkHeartbeats(responsePipe, auditTrail, configurationStorage, enumerator);
 
-                    while (await responsePipe.WaitToReadAsync(token).ConfigureAwait(false))
+                    while (await responsePipe.WaitToReadAsync(LeadershipToken).ConfigureAwait(false))
                     {
                         while (responsePipe.TryRead(out var response, out var replicator))
                         {
@@ -179,13 +169,13 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
                             {
                                 RenewLease(startTime.Elapsed);
                                 UpdateLeaderStickiness();
-                                await configurationStorage.ApplyAsync(token).ConfigureAwait(false);
+                                await configurationStorage.ApplyAsync(LeadershipToken).ConfigureAwait(false);
                             }
 
                             if (result.Value && ++commitQuorum == majority)
                             {
                                 // majority of nodes accept entries with at least one entry from the current term
-                                var count = await auditTrail.CommitAsync(currentIndex, token).ConfigureAwait(false); // commit all entries starting from the first uncommitted index to the end
+                                var count = await auditTrail.CommitAsync(currentIndex, LeadershipToken).ConfigureAwait(false); // commit all entries starting from the first uncommitted index to the end
                                 Logger.CommitSuccessful(currentIndex, count);
                             }
                         }
@@ -205,10 +195,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
                 finally
                 {
                     var broadcastTime = startTime.ElapsedMilliseconds;
-
-                    if (forced)
-                        GCSettings.LatencyMode = latencyMode;
-
+                    latencyScope.Dispose();
                     LeaderState.BroadcastTimeMeter.Record(broadcastTime, MeasurementTags);
                 }
 
@@ -216,12 +203,11 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
                 replicationQueue.Drain();
 
                 // wait for heartbeat timeout or forced replication
-                forced = await WaitForReplicationAsync(startTime, period, token).ConfigureAwait(false);
+                forced = await WaitForReplicationAsync(startTime, period, LeadershipToken).ConfigureAwait(false);
             }
         }
         finally
         {
-            cancellationSource?.Dispose();
             enumerator.Dispose();
         }
     }
@@ -247,8 +233,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
     /// <param name="period">Time period of Heartbeats.</param>
     /// <param name="transactionLog">Transaction log.</param>
     /// <param name="configurationStorage">Cluster configuration storage.</param>
-    /// <param name="token">The toke that can be used to cancel the operation.</param>
-    internal void StartLeading(TimeSpan period, IAuditTrail<IRaftLogEntry> transactionLog, IClusterConfigurationStorage configurationStorage, CancellationToken token)
+    internal void StartLeading(TimeSpan period, IAuditTrail<IRaftLogEntry> transactionLog, IClusterConfigurationStorage configurationStorage)
     {
         var members = Members;
         context = new(members.Count);
@@ -262,7 +247,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
             member.State = state;
         }
 
-        heartbeatTask = DoHeartbeats(period, transactionLog, configurationStorage, members, token);
+        heartbeatTask = DoHeartbeats(period, transactionLog, configurationStorage, members);
         LeaderState.TransitionRateMeter.Add(1, in MeasurementTags);
     }
 
@@ -270,7 +255,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
     {
         try
         {
-            timerCancellation.Cancel(false);
+            timerCancellation.Cancel(throwOnFirstException: false);
             replicationEvent.CancelSuspendedCallers(LeadershipToken);
             await (heartbeatTask ?? Task.CompletedTask).ConfigureAwait(false); // may throw OperationCanceledException
         }
@@ -298,6 +283,7 @@ internal sealed partial class LeaderState<TMember> : RaftState<TMember>
             replicationEvent.Dispose();
 
             context.Dispose();
+            heartbeatTask = null;
         }
 
         base.Dispose(disposing);