+++ CHANGELOG.md
@@ -1,6 +1,52 @@
 Release Notes
 ====
 
+# 02-28-2024
+<a href="https://www.nuget.org/packages/dotnext/5.1.0">DotNext 5.1.0</a>
+* Added `Span.Advance<T>` extension method for spans
+* `CollectionType.GetItemType` now correctly recognizes enumerable pattern even if target type doesn't implement `IEnumerable<T>`
+
+<a href="https://www.nuget.org/packages/dotnext.metaprogramming/5.1.0">DotNext.Metaprogramming 5.1.0</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.unsafe/5.1.0">DotNext.Unsafe 5.1.0</a>
+* Added `UnmanagedMemory.AsMemory` static method that allows to wrap unmanaged pointer into [Memory&lt;T&gt;](https://learn.microsoft.com/en-us/dotnet/api/system.memory-1)
+
+<a href="https://www.nuget.org/packages/dotnext.threading/5.1.0">DotNext.Threading 5.1.0</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.io/5.1.0">DotNext.IO 5.1.0</a>
+* Merged [225](https://github.com/dotnet/dotNext/pull/225)
+* Added `AsUnbufferedStream` extension method for [SafeFileHandle](https://learn.microsoft.com/en-us/dotnet/api/microsoft.win32.safehandles.safefilehandle) class
+
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/5.1.0">DotNext.Net.Cluster 5.1.0</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/5.1.0">DotNext.AspNetCore.Cluster 5.1.0</a>
+* Updated dependencies
+
+# 02-25-2024
+<a href="https://www.nuget.org/packages/dotnext/5.0.3">DotNext 5.0.3</a>
+* Fixed behavior to no-op when `GCLatencyModeScope` is initialized to default
+
+<a href="https://www.nuget.org/packages/dotnext.metaprogramming/5.0.3">DotNext.Metaprogramming 5.0.3</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.unsafe/5.0.3">DotNext.Unsafe 5.0.3</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.threading/5.0.3">DotNext.Threading 5.0.3</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.io/5.0.3">DotNext.IO 5.0.3</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/5.0.3">DotNext.Net.Cluster 5.0.3</a>
+* Attempt to fix [221](https://github.com/dotnet/dotNext/issues/221)
+
+<a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/5.0.3">DotNext.AspNetCore.Cluster 5.0.3</a>
+* Attempt to fix [221](https://github.com/dotnet/dotNext/issues/221)
+
 # 02-17-2024
 <a href="https://www.nuget.org/packages/dotnext/5.0.2">DotNext 5.0.2</a>
 * Fixed XML docs